adk-workflows/requirements.txt
google-adk>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0
